~ 7.4

+++ controllers/videoController.js
@@ -41,7 +41,10 @@ export const postUpload = async (req, res) => {
     fileUrl: path,
     title,
     description,
+    creator: req.user.id,
   });
+  req.user.videos.push(newVideo.id);
+  req.user.save();
   console.log(newVideo);
   res.redirect(routes.videoDetail(newVideo.id));
 };
@@ -51,7 +54,11 @@ export const videoDetail = async (req, res) => {
     params: { id },
   } = req;
   try {
-    const video = await Video.findById(id);
+    // populate : 객체를 데려오는 함수 (objectId 타입에만 사용 가능하다)
+    // findById(id)는 creator의 objectID만 보여주는 반면
+    // findById(id).populate("creator")는 creator객체의 정보까지 가져옴
+    const video = await Video.findById(id).populate("creator");
+    console.log(video);
     res.render("videoDetail", { pageTitle: video.title, video });
   } catch (error) {
     console.log(error);
@@ -65,7 +72,14 @@ export const getEditVideo = async (req, res) => {
   } = req;
   try {
     const video = await Video.findById(id);
-    res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+    console.log(video.creator, req.user.id, video.creator === req.user.id);
+    console.log(typeof video.creator, typeof req.user.id); // object string
+    // type이 달라서 !== 로 비교하면 true 반환함
+    if (video.creator != req.user.id) {
+      throw Error();
+    } else {
+      res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+    }
   } catch (error) {
     res.redirect(routes.home);
   }
@@ -90,7 +104,11 @@ export const deleteVideo = async (req, res) => {
   } = req;
   try {
     //Find and Delete
-    await Video.findOneAndRemove({ _id: id });
+    if (video.creator != req.user.id) {
+      throw Error();
+    } else {
+      await Video.findOneAndRemove({ _id: id });
+    }
   } catch (error) {}
   res.redirect(routes.home);
 };
